Rename path_smoothing -> path_smoother
