+++ planning/path_smoother/CMakeLists.txt
@@ -1,12 +1,12 @@
 cmake_minimum_required(VERSION 3.14)
-project(path_smoothing)
+project(path_smoother)
 
 find_package(autoware_cmake REQUIRED)
 autoware_package()
 
 find_package(Eigen3 REQUIRED)
 
-ament_auto_add_library(path_smoothing SHARED
+ament_auto_add_library(path_smoother SHARED
   # node
   src/elastic_band_smoother.cpp
   # algorithms
@@ -15,14 +15,14 @@ ament_auto_add_library(path_smoothing SHARED
   src/utils/trajectory_utils.cpp
 )
 
-target_include_directories(path_smoothing
+target_include_directories(path_smoother
   SYSTEM PUBLIC
     ${EIGEN3_INCLUDE_DIR}
 )
 
 # register node
-rclcpp_components_register_node(path_smoothing
-  PLUGIN "path_smoothing::ElasticBandSmoother"
+rclcpp_components_register_node(path_smoother
+  PLUGIN "path_smoother::ElasticBandSmoother"
   EXECUTABLE elastic_band_smoother
 )
