+++ planning/path_smoother/docs/eb.md
