+++ planning/path_smoother/include/path_smoother/utils/trajectory_utils.hpp
@@ -12,15 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef PATH_SMOOTHING__UTILS__TRAJECTORY_UTILS_HPP_
-#define PATH_SMOOTHING__UTILS__TRAJECTORY_UTILS_HPP_
+#ifndef PATH_SMOOTHER__UTILS__TRAJECTORY_UTILS_HPP_
+#define PATH_SMOOTHER__UTILS__TRAJECTORY_UTILS_HPP_
 
 #include "interpolation/linear_interpolation.hpp"
 #include "interpolation/spline_interpolation.hpp"
 #include "interpolation/spline_interpolation_points_2d.hpp"
 #include "motion_utils/trajectory/trajectory.hpp"
-#include "path_smoothing/common_structs.hpp"
-#include "path_smoothing/type_alias.hpp"
+#include "path_smoother/common_structs.hpp"
+#include "path_smoother/type_alias.hpp"
 
 #include <Eigen/Core>
 
@@ -35,7 +35,7 @@
 #include <string>
 #include <vector>
 
-namespace path_smoothing
+namespace path_smoother
 {
 namespace trajectory_utils
 {
@@ -140,7 +140,7 @@ std::optional<size_t> updateFrontPointForFix(
     motion_utils::calcSignedArcLength(points, 0, front_fix_point.pose.position);
   if (0 < lon_offset_to_prev_front) {
     RCLCPP_WARN(
-      rclcpp::get_logger("path_smoothing.trajectory_utils"),
+      rclcpp::get_logger("path_smoother.trajectory_utils"),
       "Fixed point will not be inserted due to the error during calculation.");
     return std::nullopt;
   }
@@ -151,7 +151,7 @@ std::optional<size_t> updateFrontPointForFix(
   constexpr double max_lat_error = 3.0;
   if (max_lat_error < dist) {
     RCLCPP_WARN(
-      rclcpp::get_logger("path_smoothing.trajectory_utils"),
+      rclcpp::get_logger("path_smoother.trajectory_utils"),
       "New Fixed point is too far from points %f [m]", dist);
   }
 
@@ -173,5 +173,5 @@ void insertStopPoint(
   std::vector<TrajectoryPoint> & traj_points, const geometry_msgs::msg::Pose & input_stop_pose,
   const size_t stop_seg_idx);
 }  // namespace trajectory_utils
-}  // namespace path_smoothing
-#endif  // PATH_SMOOTHING__UTILS__TRAJECTORY_UTILS_HPP_
+}  // namespace path_smoother
+#endif  // PATH_SMOOTHER__UTILS__TRAJECTORY_UTILS_HPP_
