+++ planning/path_smoother/media/debug/eb_fixed_traj_visualization.png
