+++ planning/path_smoother/media/debug/eb_traj_visualization.png
