+++ planning/path_smoother/src/utils/trajectory_utils.cpp
@@ -12,10 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "path_smoothing/utils/trajectory_utils.hpp"
+#include "path_smoother/utils/trajectory_utils.hpp"
 
 #include "motion_utils/motion_utils.hpp"
-#include "path_smoothing/utils/geometry_utils.hpp"
+#include "path_smoother/utils/geometry_utils.hpp"
 
 #include "autoware_auto_planning_msgs/msg/path_point.hpp"
 #include "autoware_auto_planning_msgs/msg/trajectory_point.hpp"
@@ -29,7 +29,7 @@
 #include <stack>
 #include <vector>
 
-namespace path_smoothing
+namespace path_smoother
 {
 namespace trajectory_utils
 {
@@ -106,4 +106,4 @@ void insertStopPoint(
   traj_points.insert(traj_points.begin() + stop_seg_idx + 1, additional_traj_point);
 }
 }  // namespace trajectory_utils
-}  // namespace path_smoothing
+}  // namespace path_smoother
